refactor: Palace 위치 변경
